Simplify to one test file
